change the server with mvc architecture
